+++ src/main/java/redis/mock/ByPassGetSet.java
@@ -4,9 +4,9 @@ import redis.CompressedValue;
 import redis.persist.OneSlot;
 
 public interface ByPassGetSet {
-    void put(byte slot, String key, int bucketIndex, CompressedValue cv);
+    void put(short slot, String key, int bucketIndex, CompressedValue cv);
 
-    boolean remove(byte slot, String key);
+    boolean remove(short slot, String key);
 
-    OneSlot.BufOrCompressedValue getBuf(byte slot, byte[] keyBytes, int bucketIndex, long keyHash);
+    OneSlot.BufOrCompressedValue getBuf(short slot, byte[] keyBytes, int bucketIndex, long keyHash);
 }
