+++ src/main/java/redis/repl/incremental/XChunkSegmentFlagUpdate.java
@@ -77,7 +77,7 @@ public class XChunkSegmentFlagUpdate implements BinlogContent {
 
 
     @Override
-    public void apply(byte slot, ReplPair replPair) {
+    public void apply(short slot, ReplPair replPair) {
         var oneSlot = localPersist.oneSlot(slot);
 
         for (var entry : updatedChunkSegmentFlagWithSeqMap.entrySet()) {
