+++ src/main/java/redis/repl/incremental/XDict.java
@@ -93,7 +93,7 @@ public class XDict implements BinlogContent {
     }
 
     @Override
-    public void apply(byte slot, ReplPair replPair) {
+    public void apply(short slot, ReplPair replPair) {
         log.warn("Repl slave get dict, key prefix or suffix: {}, seq: {}", keyPrefixOrSuffix, dict.getSeq());
         // ignore slot, need sync
         var dictMap = DictMap.getInstance();
