+++ src/test/groovy/redis/command/HGroupTest.groovy
@@ -172,7 +172,7 @@ class HGroupTest extends Specification {
         reply == NilReply.INSTANCE
     }
 
-    final byte slot = 0
+    final short slot = 0
 
     def 'test prefer use hh'() {
         given:
